requirements.txt
Django==4.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
django-phone-field==1.8.1
Pillow==9.0.2
djangorestframework-simplejwt==5.2.0